Fix x86 32-bit wrong variable type for rdtsc
Bruce points out, that we are putting the result into a 32-bit
variable. That is incorrect, make it a proper ULL instead.

+++ arch/arch-x86.h
@@ -34,7 +34,7 @@ static inline unsigned long arch_ffz(unsigned long bitmask)
 
 static inline unsigned long long get_cpu_clock(void)
 {
-	unsigned long ret;
+	unsigned long long ret;
 
 	__asm__ __volatile__("rdtsc" : "=A" (ret));
 	return ret;
